changed SaveSnippet to SaveSnippetCopy, removed Modal code from SingleSnippet
